refactoring and cleaning up the code to make it dry
