+++ package.json
@@ -13,6 +13,7 @@
     "react-icons": "^4.7.1",
     "react-router-dom": "^6.8.1",
     "react-scripts": "5.0.1",
+    "react-uuid": "^2.0.0",
     "sweetalert2": "^11.7.1",
     "web-vitals": "^2.1.4"
   },
