+++ src/Components/Footer.js
@@ -1,0 +1,17 @@
+const Footer = () => {
+  return (
+    <footer>
+        <p>
+          Made by <a href="http://www.kylemoore.ca" target="_blank" rel="noreferrer">
+            Kyle
+          </a>, <a href="http://umeshcodes.com/" target="_blank" rel="noreferrer">
+            Umesh
+          </a>, and <a href="https://vincentyoung.dev/" target="_blank" rel="noreferrer">
+            Vincent
+          </a> at Juno College
+        </p>
+      </footer>
+  )
+}
+
+export default Footer
