+++ src/Components/Home.js
@@ -2,17 +2,14 @@ import Form from "./Form";
 import Results from "./Results";
 import Loader from "./Loader";
 
-const Home = ({ apiCall, wordList, isLoading }) => {
+const Home = ({ wordList, isLoading, setIsLoading, setWordList }) => {
   return (
     <>
-      <Form apiCall={apiCall} />
+      <Form setWordList={setWordList} setIsLoading={setIsLoading} />
       <Loader isLoading={isLoading} />
-      
       {
         (wordList.length === 0) ? null :
         <Results wordList={wordList} />
-      
-      
       }
     </>
   );
