+++ src/Components/Loader.js
@@ -1,4 +1,5 @@
 import favicon from "../assets/favicon.png";
+
 const Loader = ({ isLoading }) => {
   const loaderClass = isLoading ? "loader loading" : "loader";
   return (
