+++ src/modules/firebase.js
@@ -1,9 +1,5 @@
-// Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
 
-// Your web app's Firebase configuration
 const firebaseConfig = {
   apiKey: "AIzaSyCR1-jTMmp5MCPmYgTxkC9M_rneJf0R82Q",
   authDomain: "word-party-7947f.firebaseapp.com",
@@ -14,6 +10,6 @@ const firebaseConfig = {
   appId: "1:703213887524:web:47005cb362841c90c4f8b1"
 };
 
-// Initialize Firebase
 const firebase = initializeApp(firebaseConfig);
+
 export default firebase;
